requirements.txt
flask==3.0.3
python-dotenv==1.0.1
PyJWT[crypto]==2.9.0
requests==2.32.3
openai==1.100.2
pytest==8.3.3
pytest-timeout==2.3.1
